+++ src/Steps/Login.vue
@@ -6,12 +6,18 @@
     >
       <v-flex v-if="notLoggedIn">
         <LoginForm
-          v-if="!createAccount"
+          v-if="!createAccount && !forgotPassword"
           @createAccount="toggleCreateAccount"
+          @forgotPassword="toggleForgotPassword"
         />
         <CreateUserForm
-          v-else
+          v-else-if="createAccount && !forgotPassword"
           @login="toggleCreateAccount"
+        />
+        <ForgotPasswordForm
+          v-else
+          @login="toggleForgotPassword"
+          @sendRequest="handleSendRequest"
         />
       </v-flex>
       <v-flex v-else>
@@ -31,6 +37,21 @@
           </v-card-text>
         </v-card>
       </v-flex>
+      
+      <v-snackbar
+        v-model="snackAlert"
+        :color="color"
+        :timeout="timeout"
+      >
+        {{ text }}
+        <v-btn
+          color="white"
+          text
+          @click="snackAlert = false"
+        >
+          Close
+        </v-btn>
+      </v-snackbar>
     </v-layout>
   </v-container>
 </template>
@@ -45,20 +66,28 @@
 import _ from 'lodash';
 import LoginForm from '../Components/Authentication/LoginForm.vue';
 import CreateUserForm from '../Components/Authentication/CreateUserForm.vue';
+import ForgotPasswordForm from '../Components/Authentication/ForgotPasswordForm.vue';
 
 export default {
   name: 'Login',
 
   components: {
     LoginForm,
-    CreateUserForm
+    CreateUserForm,
+    ForgotPasswordForm,
   },
 
   data: () => ({
     username: '',
     password: '',
     error: '',
-    createAccount: false
+    color: '#0abb8a',
+    createAccount: false,
+    forgotPassword: false,
+    snackAlert: false,
+    timeout: 3000,
+    text: 'Reset email has been sent',
+
   }),
 
   computed: {
@@ -95,6 +124,13 @@ export default {
     toggleCreateAccount() {
       this.createAccount = !this.createAccount;
     },
+    toggleForgotPassword() {
+      this.forgotPassword = !this.forgotPassword;
+    },
+    handleSendRequest() {
+      this.forgotPassword = !this.forgotPassword;
+      this.snackAlert = true;
+    },
     /**
      * clear the store on logout.
      */
@@ -105,5 +141,3 @@ export default {
 
 };
 </script>
-
-