+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [0.1.40] = 2020-4-10
+### Changed
+- Added email on signup form
+- Finished password reset
 
 ## [0.1.39] = 2020-4-9
 ### Changed